In NQueen.Kernel   - Changed the following types from byte to int: BoardSize, row and column number, as well as all related fields, variables, constants, and parameters   - Added an Excel Sheet for simulation resultsto the solution

+++ NQueen.Console/Commands/DispatchCommands.cs
@@ -5,7 +5,7 @@ public partial class DispatchCommands(
     IConsoleUtils consoleUtils,
     ICommandProcessor commandProcessor)
 {
-    public byte BoardSize { get; set; }
+    public int BoardSize { get; set; }
 
     public SolutionMode SolutionMode { get; set; }
